requirements.txt
flask>=2.0.0
flask-sqlalchemy>=3.0.0
nemoguardrails>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
cryptography>=41.0.0
regex>=2023.0.0
spacy>=3.6.0
psycopg2-binary>=2.9.5
writer-sdk>=1.0.0 